+++ error.c
@@ -243,7 +243,7 @@ noreturn static pic_value
 pic_error_error(pic_state *pic)
 {
   const char *str;
-  size_t argc;
+  int argc;
   pic_value *argv;
 
   pic_get_args(pic, "z*", &str, &argc, &argv);
@@ -257,7 +257,7 @@ pic_error_make_error_object(pic_state *pic)
   struct pic_error *e;
   pic_sym type;
   pic_str *msg;
-  size_t argc;
+  int argc;
   pic_value *argv;
 
   pic_get_args(pic, "ms*", &type, &msg, &argc, &argv);
